VIC-3317: Build flag ANKI_NO_WEBSERVER_ENABLED should prevent deploy of webserver resources (#2169)
The cmake used to manage resource deployment was getting hard to manage.  I added a helper macro add_asset_target()
to simplify the task of managing deployment targets.  Each target is automatically added to the global ASSET_TARGETS, and it gets a dependency on the last
target in the sequence.  This makes it easier to add/remove targets without breaking the chain of dependencies.

Net change: Building with FACTORY_TEST=1 or ANKI_NO_WEBSERVER_ENABLED=1 will prevent deployment of webserver resources.

Bonus fix: Calling victor_build.sh with '-F feature' now forces configuration. This was an oversight when I added '-F'.