VIC-3611 Come here uses mic direction (#2151)
Adds an option to the behavior for the Come Here voice command to
know if it started from a voice command, in which case it can assume
it activates facing toward the dominant mic direction. If no face is
seen in the mic direction but a previous one was known, it turns
back toward that. If any face can be verified, it then drives toward
it. If not, it allows for search behavior to find one before driving
toward it.