VIC-3910: Minimal script to upload robot logs to blobstore (#2224)
1. New script 'vic-logmgr-upload' runs on robot to collect a snapshot of log data and upload to blobstore.
The upload is tagged with ESN, timestamp, and version fields to help identify the file.

2. New wrapper script 'victor_log_upload.sh' can be run by QA to invoke (1) on robot.

Blobstore requires an application key for each file submitted. SAI will issue an application key for this usage shortly.

Note the application key is not stored on the robot. Uploads can only be performed by users with access to internal development script.