VIC-3612/VIC-3613/VIC-3609 VCs: Look at me, What's my name, Yes (#2276)
For the "Come Here" VC, BehaviorDriveToFace was recently expanded to
look for faces in multiple places (mic direction and previous known
poses) before driving to the face.
This splits that behavior into its prior form as BehaviorDriveToFace and
a new BehaviorFindFaceAndThen. The latter finds the face, and if found,
it delegates to the former.
VIC-3612/VIC-3613 New VCs "look at me" and "what's my name" use
this pattern.
VIC-3609 the affirmative voice command.