Add message to notify engine when streaming starts (#2298)
Currently, the engine is having difficulty (in the absence of any
response from the cloud) identifying the cause of a timeout. A
delay of several seconds could be due to...
- the user not saying anything, or
- a connection error (which can sometimes take dozens of seconds
  to be returned, depending on the cause)

To help clear this up, this change adds a message sent from the
cloud process to the engine as soon as a connection is successfully
made with the server after a trigger word. This will allow the
engine to infer the cause of a timeout:
- if it received the "connection successful" event and a timeout
  occurred, it was probably due to the user not speaking
- if it did not receive the "connection successful" event and a
  timeout occurred, it's likely to be a connection error