VIC-1539 VIC-876 Remove 'audioName' data from audio keyframes (#2320)
When converting animation to binary format, this will now check every event group and remove the 'audioName' data if present. We like the 'audioName' field in JSON data for humans, but the engine only uses an event's numerical ID, so we strip out the name string before converting to binary format.