VIC-4036: Pairing screen disappears (#2326)
* VIC-4036: Pairing screen disappears shortly after double press
This only happened after shaking (i.e. when robot is engaged in regular behaviors / playing animations). It seemed like engine was cancelling a PlayAnimationAction which sends AbortAnimation to anim process which aborts the currently running animation which is the pairing screen image. AbortAnimation should only cancel the animation that the engine had commanded.

* Fixes possible race condition whereby animation process could play the pairing screen animation and have it immediately cancelled by an animation from engine. By making engine go to BehaviorWait based on the switchboard SetConnectionStatus message, it's behaviors should now already be cancelled by the time anim process goes to the pairing screen.

* AnimationStreamer::Abort() wasn't calling StopTracksInUse() which was causing robot to sometimes drive wheels when going into debug screen.