VIC-3202 Use systemd for automatic & manual OTA updates
We want Victor to periodically try to perform an OTA update.  In
addition, we want to preserve the ability of the end user to do a manual
update with Chewie to a specific URL.  With this change, we rely on
systemd to make sure that 1 and only 1 instance of
/anki/bin/update-engine is running.  Furthermore, we will use a systemd
service file for update-engine to make sure that the necessary
preconditions are met before an update is started.  This systemd service
file will also set environment variables for update-engine to use during
its execution.

This change to update-engine also supports sharding when requesting an
OTA update.  However, it is disabled by default.

This change to vic-switchboard and update-engine will also run on
versions of the Victor OS that DO NOT have the update-engine systemd
service.