Add cloud token client (#2197)
* Add JWT token store

The token store (package anki/token/jwt) is responsible for:
- parsing encoded tokens passed in (presumably sent to us by the
  server)
- persisting tokens on disk
- returning the currently stored token when asked

* Add token client

The token client (package anki/token) is responsible for handling
requests for tokens to use for authorization against Anki cloud
services. It runs a domain socket server for external processes
to send token requests, or it can be called directly from code
using it (as the cloud process does). Its primary duties are:
- When asked, perform a primary user authentication with the Token
  Management Service (TMS), passing a session token which will be
  handed to the token client from the onboarding process; this
  stores the initial JWT token, which has 24 hour expiration and
  will be used for authenticated requests to Anki cloud services,
  like Chipper
- When asked, return the current JWT web token if it's still valid;
  if it requires refreshing, it will silently ask the server to
  refresh the token and return the updated token

To avoid issues with multiple requests taking place at the same
time, all requests are put onto a single request queue which is
processed on a dedicated go routine. The routine will attempt to
re-use existing RPC connections, but in practice most requests
will probably not coincide with other requests.

* Update cloud process to use new token package

For voice requests, the cloud process will now ask the token
package for auth tokens. For the time being, if one can't be
obtained, it will fake an "associate primary user" operation to
get a new token.