For China: don't fail voice requests if we can't get a token (#2373)
To make failures easier to spot, the robot client will fail voice requests if it
can't retrieve a token from the TMS service. This has worked consistently in the
office, but appears to fail in China, so to prevent that problem from blocking
people, this change will continue with a voice request even if a token can't be
obtained.