VIC-4173 Fix neutral eyes race condition when exiting dev screens (#2376)
When entering the pairing screen or CC screen, facekeepalive was
disabled, but "neutral eyes" persisted. When exiting those screens, the
engine resets its behavior stack and may or may not send an animation.
We need the keepalive and neutral eyes in case an animation is not sent,
but the engine needs enough time to send it before they are shown.
Otherwise, neutral eyes momentarily flash before the engine's animation
begins.
This makes the toggle for KeepFaceAlive also apply to "neutral eyes."
It also removes some KeepFaceAlive(true) calls, replacing them with a
message from the engine that is sent a few ticks after it resets the
behavior stack.