VIC-4174 Touch /run/on-charge-contacts
For automatic over the air updates, we want to know if Victor is on his
charger or if he is just using battery.  With this change, `vic-engine`
will touch /run/on-charge-contacts when the robot is on the charging
contacts and will remove /run/on-charge-contacts when the robot is off
the charging contacts (and just using battery).