CLAI-200 Fix cloud process blocking on upload issues (#2363)
Ross encountered an issue on his home network where the cloud
process appeared to stop uploading audio, and a few seconds later
the `vic-anim` log started filling with "sent -1 bytes" errors.
The cloud process later returned a server error of EOF.

My best guess for what happened is:
- the upload encountered a connection problem and hung
- the audio channel, which is buffered, filled up with more audio
  packets to send
- when the audio channel filled, the main cloud process routine
  blocked trying to put more audio into it, while the GRPC stream
  was still blocked on an upload
- after a few seconds, the domain socket buffer fills and
  `vic-anim` is no longer able to send into it

To fix this issue, this change refactors the way audio data is
buffered to upload via GRPC. The new code adds a goroutine that
sits between the main routine and the streaming routine to make
sure the main routine cannot be blocked on the GRPC routine. The
new routine uses a select statement to pull audio data from the
main routine or push it to the upload routine when either one is
ready; if the upload routine is blocked, incoming audio will still
be promptly pulled from the main routine and be added to a buffer
that can grow however large is needed.