VIC-4059 Wake word turn animations not turning (#2329)
Fixing BodyMotionKeyFrame::IsDone() so it allows a keyframe to play if it is the only BodyMotionKeyFrame in the track and it has a trigger time of 0 ms. Specifically, updating that to automatically returns false if the trigger time is equal to the time since the animation started, i.e. if it is now time to play that keyframe.