Simplify the charger docking Webots test (#2399)
Consider the test to have passed if we sense the charger contacts at any time. Sometimes due to Webots physics inconsistencies we hit the contacts then overshoot or shift to the side. For the purposes of this test (which is to ensure that we find the charger and at least attempt to dock with it), we only care if the robot makes it to the charger contacts.