VIC-3889 Fix robot waking up during pairing (#2407)
If the pairing process gets canceled, perhaps because the backpack
button wasn't double-pressed, then the connection flow manager
receives a message with an END_PAIRING status. This aborts the
current animation. If pairing didn't actually start, then that
animation could be part of a behavior that is still running. In
this case, the animation was from a behavior that wakes up after
its first animation ends.
The fix checks to see if any pairing screens have been drawn.
This also removes toggles to a console var that no longer exists.