VIC-4177 cmake-based asset copy doesn't work for some people (#2381)
Adds a cmake option DEPLOY_WITH_CMAKE, this is enabled by default.

When DEPLOY_WITH_CMAKE is enabled cmake.exe is used for copying
assets into the data directory. This works for most people and
team city.

When DEPLOY_WITH_CMAKE is disabled, instead of spawning cmake.exe
python is run instead with a script that replicates the cmake
functionality. This works for at least a person.