VIC-3366 Cube spoonbender (#2404)
Hooks up animations for cube activation ("spoonbending").
This involves a new behavior that turns towards the cube, moves
forward or backward if poorly spaced, and plays an animation.
Stage logic is updated to run this prior to the pickup behavior,
but not block if it for some reason can't see the cube during this
stage. The cube was already seen once, and it can still be
"activated" even if the robot isn't facing it.