VIC-4054 Photos stop being stored to robot (#2444)
* Prevent PhotographyManager from getting stuck when camera daemon restarts

If the camera daemon randomly restarts while waiting for a format switch, PhotographyManager would get stuck in a weird state waiting for an image to be saved in a certain format. This adds a function to CancelTakePhoto when it has taken too long to actually TakePhoto

* Fix race condition when pausing VisionComponent

Fixes race condition when pausing VisionComponent right after _nextImg has been swapped in _currentImg but before UpdateVisionSystem has been called. This was preventing _currentImg from being cleared which caused VisionComponent to get stuck WaitingForProcessingToStop when switching camera formats

* Fix crash when switching formats while converting YUV images

Need to make sure we are not actively converting an image when switching camera formats since that image memory will be deallocated during the format switch

* Protect against bad data in camera_frame_acquire

* Finish converting YUV image before switching formats

Need to let image conversion finish before switching formats otherwise we will be using invalid memory

* Don't set exposure/gain or white balance while waiting for a format change