VIC-4226 Robot should poke at things during exploring (#2425)
If the robot has scanned some object during exploring, and the object
is of a nominal size, it now will try to push it. First, it will
push at a low speed and then retreat. If the robot is still centered
well, it will try again at high speed. If the robot is evil, it will
try to push things off cliffs.
A future commit will add anims