VIC-1868 Redo status messaging (#2467)
Status was a CLAD union for all things that should go in the status log
in the app. In order to send a status history, which is an array of
recent statuses, all elements of the array must be of the same type.
This means we can't have both CLAD statuses and proto statuses.
This replaces all Status unions with proto Status oneofs. It also
creates RobotHistoryRequest/RobotHistoryResult messages.
Event is currently not used on the CLAD side, but I left it there in
case someone really can't use protobuf for something.