CLAI-202 Auto-refresh auth tokens when available
This change adds a routine that will run in the background,
checking for when we have a valid token (initially, until full
token request flow is hooked up, it will also submit a fake request
to obtain an initial one from the server), and sleeping until it's
time to refresh the token (3 hours before expiration).

One edge case I may want to take into consideration in the future
is that if someone happens to be using voice services at the exact
moment 21 hours later that the background routine decdies to
refresh the token, the existing token will become invalid.

This also cleans up the token code a little bit (allows its
goroutines to be stopped, sets user agent on all requests).

Since PVT cert issues should be fixed now, this also reverts the
change that allowed voice requests to continue even if obtaining a
token failed.