VIC-3472 Hotfix to prevent unnecessary name change (#2534)
V2 and V3 of saved session keys file ought to be treated as the same. Accidentally, V2 was treated as old. In reality, there are always two cases: 1) file is current, good to go, or 2) file is old--parse it with old parser and handle it. Since V2 and V3 are the same in terms of session file format, it should have been case 1), but this was accidentally overlooked in the christening case.