VIC-3996: Victor-side changes for new vic-on-failure service (#2447)
Application-level scripts to support coordinated restart of services.
Corresponding service changes will be posted in a separate PR.

The general idea is that on failure, vic-on-failure will call /bin/animfail and/or restart services,
depending on configuration.  This will make it easier to adjust configuration for release vs demo vs shipping behaviors.