VIC-3943 Fix turn away from player in BlackJack (#2566)
* VIC-3943 Fix turn away from player in BlackJack

BlackJack reorients victor toward the player at intervals throughout
the game. Previously this was done using the
TurnTowardsLastFacePoseAction, but this was resulting in some wierd
errors likely due to an overly tolerant face-age policy. This has been
replaced with the more intelligent BehaviorLookAtFaceInFront.

Reorienting toward the face also used to happen on each hit/stand
prompt, but this seemed like overkill/occasionally ruined the
experience when it didn't work well. The prompt behavior is
reconfigured to NOT look at a face before prompting by default any
more, and the surrounding logic should probably be re-evaluated at a
later time.

* Add missing wantsToBeActivated check