VIC-4289 proper reaction to body (#2560)
New behavior that reacts when a body is detected. The procedure is as follow:

If a face is already detected, nothing to do(play animation? sing a song?)
If on charger, drive off
If configured drive straight a bit
Delegate to SearchInBoundingBox using the SalientPoint extents as the search area
If a face is detected, stop the search (and play animation? sing a song?)