Add referencing behavior prior to exploring poke (#2600)
If the robot is about to poke an object, it now sometimes will
reference a person before that.
For the other case where this behavior is run, I also replaced the
cooldown with a probability so that they could share this same
instance.

(cherry picked from commit da8ecb7a6d5136cfa8a326301a1ec79be129c1dc)