VIC-4677 Fix CubeConnectionCoordinator callbacks (#2608)
The `iCubeConnectionSubscriber::Connected_Callback` functions were only
invoked when the connection first transitions to the applicable state,
so, if a subscriber requests a connection which is already active, the
callback will never be called for that subscriber.

These changes collapse the Interactable and Background callbacks to a
single callback with an ECubeConnectionType param to simplify the API
and calling structure, and invoke that callback immediately upon
subscription if a cube connection is already active. Implementers
should then handle the callback making account of the connection type.