VIC-4661 switchboard : Fix failure to restart OTA
This looks like a C++ class inheritance problem.  `IRtsHandler`, the
base class, had 2 private members : `_isPairing` and `_isOtaUpdating`.
There are 2 public setters : `SetIsPairing` and `SetOtaUpdating`.
However, the derived classes, `RtsHandlerV2` and `RtsHandlerV3` also had
members named `_isPairing` and `_isOtaUpdating`.  Other code like the
`Daemon` class would call into `RtsComms` to change the state of the
`_isPairing` and `_isOtaUpdating` variables.  `RtsComms` in turn would
call into `SetIsPairing` and `SetOtaUpdating` in `IRtsHandler`.
However, code in `RtsHandlerV2` or `RtsHandlerV3` would rely on the
values of its members by the same name.  Confused yet?

To fix this, I removed the `_isPairing` and `_isOtaUpdating` members of
`RtsHandlerV2` and `RtsHandlerV3` and instead created a protected
`IRtsHandler` constructor for them to both call in their own
constructors.  Now, when `Daemon` calls into `RtsComms` to change the
state of these variables, it is seen in the derived classes.