CLAI-213 Send language in hotword message (#2623)
Since the trigger word model needs to be set to a given language
for us to use it anyway, this change has `MicDataSystem` store the
locale that the trigger word model is set to and sends it up to
the cloud client when making a hotword request. The cloud client
will then use the locale string to determine the language and
include it in server requests.

github.com/anki/sai-chipper-voice:
a3e8cc7 Add language as option in client

(cherry picked from commit 551772aadd885731eba28c41f266162c62d6f86b)