VIC-4725 Reflect navmap webviz (#2629)
When making the navmap webviz tab, I didn't bother changing from
positive y being up to graphics-land positive-y being down, and just
added a NOTE on the tab to tell people to use their imagination.
Turns out this is in the PR demo.
The fix is to reflect everything across an arbitrary x axis. I only
tested 2D since that's what will be shown. 3D might be broken now.
This also prevent zooming too far or close, and updates the robot and
cube images to be vector instead of cozmo. It also decreases the
update frequency.