VIC-4263 Hide planning times with animations (#2445)
Even though the planner is super speedy, we still need something to
cover the moments when it is planning or replanning. This adds a
planning get-in/loop/get-out to the DrivingAnimationHandler with the
same logic as driving animations. The planning get-in starts whenever
the robot doesn't have a plan. If a plan becomes ready by the time the
get-in finishes, it switches directly to the get-out. Otherwise, the
looping animation fills the gap until the plan is ready, at which point
the get-out plays, and then driving starts.

Changes elsewhere:
+ DriveToPoseAction optionally plays planning animations.
+ PathComponent can replan without immediately starting to drive the
  new path, so that if a planning animation has already started, we can
  end it before starting on the new path.
+ XYPlanner uses the artificial planning latency console var
+ XYPlanner updated to finish driving the current path segment, if
  it is safe. Usually the robot still stops immediately, perhaps
  because the current path segment is long.