VIC-4752 "Come here" should resume from cliff with an animation (#2654)
Because "come here" is really a BehaviorFindFaceAndThen that delegates to a BehaviorDriveToFace, the child drive to face class couldn't specify WantsToBeActivated conditions that would allow it to resume following a cliff. That would have to be done in the parent behavior. So this commit both inverts the BehaviorFindFaceAndThen/BehaviorDriveToFace relationship, so that BehaviorDriveToFace delegates to BehaviorFindFaceAndThen, and then moves the respondToUserIntents params inside of the class and adds logic to resume with an animation following a cliff strike.

(cherry picked from commit 4fdc103be4aa35a19926b6e02b9cc20e7460ba6e)