VIC-4496 Dont turn quickly after meet victor or come here VCs (#2687)
These fix issues where the robot turns toward the mic direction, then
immediately turns elsewhere.

Meet victor:
This increases the duration of the state WaitingInPlaceForFace, and
transitions to LookingForFace (which turns toward a face if known) the
moment a face is seen.

Come here:
This is only an incremental improvement for a specific edge case, but
it's a low risk change.
The only way I can repro this so far is if the robot already sees a
face, turns toward the mic direction, but still sees the original face.
Before it sees a new face directly in front of it, it turns toward the
original face. BehaviorFindFaceAndThen used to immediately turn toward
any face. Now, it holds the pose facing the mic direction for a minimum
time (0.5s) before turning toward the face. This hopefully means that
it will see the face in the mic direction, which would be newer than
the face in the periphery.
This doesn't use face selection criteria, and so isn't ideal, and it
may still choose the most recently seen face instead of the one in
front. That's a change for post-demo (or maybe tomorrow).

(cherry picked from commit 415e784922d136cd25d1fe1619de8de57d74a868)