CLAI-221 Move blocking connection operation off of main routine (#2709)
Previous issues with the socket buffer between the anim process
and cloud process filling up have been fixed by preventing the
main routine from blocking when audio is added to the stream queue
after a connection is established. However, that fix did not
address the case where the connection process itself can block for
long periods of time.

Because the code was starting to feel a little clunky, this change
first moves what was previously the `voiceContext` struct (which
handled an individual voice request) to a sub-package. Then:
- the process of establishing a connection was moved from taking
  place in the main voice process routine, where it could block,
  to instead be the responsibility of the new `Stream` (formerly
  `voiceContext` object
- that connection establishment is moved to a goroutine
- the buffer routine (to buffer incoming audio requests) is started
  before connection establishment instead of after, so that incoming
  audio messages will still be received and put on the queue even if
  the connection hasn't yet been made