VIC-3969/4054/4077/4489 "Fix" all the camera problems (#2652)
Giving up on getting camera format switching to work 100% of the time. The camera problems stem from preview mode both when switching to and from it as well as sometimes just staying in it. The solution to these problems is to just not use preview mode. Instead we now stream the raw bayer images to engine and let engine deal with if it wants to downsample or demosaic the image in order to get 640x360 or 1280x720 images.

Added new EnableFullRes function that will automatically switch the camera to the raw format, if not already in it, and set a bool that controls if we do downsampling of demosaicing. The potential format switch shouldn't be an issue as the camera is not actually switching to a different stream, just stopping and starting the raw stream and reallocating some memory. The camera will either start up in RawRGB or BAYER depending on the OS version. PhotographyManager now uses this when taking a picture.

Actually switching formats is still supported but not guaranteed to work.

There is also the added support for the snapshot format because I hooked it up to test it and may be useful in some future...

* Don't set exposure/gain or white balance while waiting for a format change

* Hooking up snapshot mode

* Support for raw Bayer images from camera

Adds ability to request raw Bayer formatted images from the camera
Adds ability to pick whether or not to downsample or demosaic the bayer image