VIC-4924 Passive habitat detection in onboarding (#2807)
We want to pop a modal when people try to start onboarding with the
habitat, which could be expected given that habitats ship for
kickstarter supporters.
The ConfirmHabitat behavior would derail the onboarding flow, so
this takes a passive approach to detecting the habitat. The robot
begins onboarding by driving off the charger and then playing an
animated response to the "come here" voice command. As luck would
have it, those two driving segments will almost surely bring the
robot to the white habitat stripe, facing outwards. When any of
the front cliff sensors see white at the right time, and the prox
sensor reading is within a specific range, this triggers a reaction
and sends a message to the app.