VIC-4407 Strongly typed timestamps (#2824)
We have clocks in different processes that provide timestamps of
the same integral type TimeStamp_t. Since the unintended use of the
wrong clock's timestamp has caused a ubiquitous myriad plethora of
bugs, this creates a class for strongly-typed POD data and uses it
for a RobotTimeStamp_t as well as an EngineTimeStamp_t and an
AnimTimeStamp_t, all of which have underlying type TimeStamp_t.
In the engine process, this replaces TimeStamp_t with the timestamp
type corresponding to process that generated the data. So anything
that originated as a message from the robot is now a RobotTimeStamp_t,
and anything that originated from an engine call
to BaseStationTimer::getInstance()->GetCurrentTimeStamp() is now an
EngineTimeStamp_t.
The anim process rarely uses robot timestamps or BaseStationTimer,
but in the few places they're used, they're now RobotTimeStamp_t and
AnimTimeStamp_t. Other timestamps are relative to the animation start
time. I debated making a TimeSinceAnimStart_t, which might actually be
a useful identifier in the engine process for things like animation
actions and track locking logic. But I didn't make that change since
the anim process is already pretty consistent.
Coretech remains entirely as TimeStamp_t, so mistakes can still be made
if we call a coretech method like GetLastObservedTime and don't
immediately cast to RobotTimeStamp_t. CLAD messages are also still
TimeStamp_t by necessity. Also, some difference variables, e.g.,
(robotTimeStampA - robotTimeStampB), are passed around as TimeStamp_t.
A generic Duration_t might be useful someday.

This also fixes some timestamp-related bugs, and makes a few changes
for consistency. For example, all cube accel-based messages now use the
robot timestamp, whereas they used to be a mix of engine and robot
timestamps.