VIC-4418 _Running webots sdk tests (#2735)
VIC-4418: Cleaned up and Edited sdkTest.py to have it run for Vector. sdkTest.py would run the tests in the worlds specified for that test in sdkNightlyTests.cfg.

Building the app has been removed from sdkTest.py as it will be build on Teamcity.
Tests can be added to sdkNightlyTests.cfg by giving it a test name, the script that needs to be run for the test, and the world in which the test needs to be run
sdk.wbt has been modified to have it open Vector as the simulated robot and have it open with its behavior system turned on.