VIC-4290 Add person detection to BehaviorFindFaceAndThen (#2839)
FindFaceAndThen spends some time looking for faces in mic direction,
and if nothing is found there, it looks wherever a face was last seen.
If no face is found, it eventually performs a full search. This adds
person (body) detection to that, so that if a body is seen in any
search direction, the robot spends a little longer facing
that direction, and runs SearchWithinBoundingBox to find a face from
the body.
This also switches BehaviorAestheticallyCenterFaces over to use
BehaviorFindFaceAndThen as a search behavior. Since the latter has its
own (dynamic) timeout, some delegation logic was changed.