VIC-4678 CubeInteractionTracker (#2885)
Adds a new EngineComponent to monitor visually and via cube connection
(when available) for physical user interaction with vector's cube.
This component is then used to supply data to a beiCondition in HLAI
which allows the user to "get vector's attention" by holding a cube in
front of him during most freeplay behaviors.

The resultant behavior, InvestigateHeldCube, tracks the cube the user
is holding via pan and tilt and monitors for conditions which indicate
useful delegations it might make. The only active delegation as of now
is Keepaway as a test case for this mode of interaction.