VIC-5141 CloseSend on stream when mic is done streaming (#2912)
If the VAD on our third party providers is unable to detect the
end of audio streaming in the data we send before the mic cuts off,
it never will. This change makes sure that when the mic is done
streaming audio to the cloud process, we call CloseSend() on the
server stream, which signals to Chipper that we're done sending
audio and it should deliver us a response ASAP. This should reduce
timeouts from the server.