VIC-5087 Configure automatic OTA updates
We want to automatically update the robot to the latest and
greatest. This change brings in a configuration file for development
builds and a separate one for shipping builds.  The configuration file
will end up on the robot in `/anki/etc/update-engine.env`.

For developers and QA who wish to disable automatic updates, they can
create the file `/data/etc/update-engine.env` with a single line:
UPDATE_ENGINE_ENABLED=False