VIC-5150 Action to reselect animations when looping (#2921)
When creating looping animations, a single animation is selected from
a group and then looped. If we want to choose a new animation each
time, a new action has to be made. In meet victor, that's not possible,
because the animation loop action runs in parallel with a track face
action. The change to have an animation trigger action reselect a new
animation each loop is not trivial.
This creates a new action that repeatedly plays new instances of a
TriggerLiftSafeAnimationAction up to some number of max loops, or
forever if numLoops==0.