CLAI-207 Add tool to authorize robots with Anki accounts (#2902)
* Step 1: Program to log in and obtain accounts session token

* Step 2: Add dev server on robot to auth w/ token

(and code in the provisioning program to connect to it)

* Refactor dev HTTP server handlers

We still want to replace an existing JWT token if the user ID on
it is nil. We also don't really want to expose the user ID (which
has no other reason to be transmitted) across CLAD requests for
tokens. And there's not a good reason for the `cloudproc` package
to be responsible for making a bunch of token requests anyway, let
alone adding code there to parse the returned JWT token for a user
ID.

So, this commit makes individual modules responsible for adding
whatever HTTP handlers they want to the dev server. That way, the
token package can be responsible for the dev authorization endpoint
and tell on its own whether the existing token is valid, or should
be updated with a real token associated with a real account.