VIC-4599 Power on boot animation (#2879)
Adds a vic-bootAnim executable which endlessly draws the boot animation to the face until it is stopped. Currently doesn't run, needs to be hooked up to the early-anim service in the OS. The plan is to have vic-anim mark early-anim as a conflict via systemd. When vic-anim starts it will cause early-anim to be stopped. This will stop the boot animation and clear the screen. The screen will stay blank until everything is fully loaded and we can start showing eyes.