BI-1238 tread tests (#2846)
Tests if the Victor's encoders are working properly, making sure that the robot is moving the amount the encoders say it is. 

The test checks if the actual distance matches the distance moved, both forward and backward. 

Checking the distance moved is done by using a robot cube as a marker, and a set origin point for Victor.