VIC-4318 Add robot battery countdown to app BatteryStateResponse (#2953)
Preflight check shows a battery countdown. This adds a remaining charge
time (in seconds) to BatteryStateResponse. It will be 0 if the robot
boots and is not low battery, which will likely happen if the robot was
on the charger from the 0.9 to 1.0 OTA. Otherwise, if it boots into low
battery, or enters a low battery state, this will contain a suggested
charge time. If the robot is removed from the charger, the countdown
pauses, and then when placed back on the charger, the time resets based
on how long the robot spent off the charger