VIC-2272 webviz communication optimizations (#2966)
* VIC-2272 webviz communication optimizations

This optimizes CPU for sending data from engine and anim, to webviz,
basically by avoiding unnecessary work when appropriate.  Brad had
made a "IsWebVizClientSubscribed" call that checks to see if there are
any listeners for a particular webviz module, and had used that in a
couple of places.  This change basically covers all the other places.

- In general, in about 16 or 20 places in the code, check to see if there
  are any listeners (with IsWebVizClientSubscribed) before doing a
  bunch of work to create data and attempt to send, only to not actually
  send the data because there were no listeners
- Also use 'static' in a few places to avoid unnecessary reinitialization
- Also use 'const &' in at least one place to avoid a string copy
- Note: There were a couple of places where not much work was being
  done, so I did not insert the call there.

Before-and-after performance measurements:

For each of 'before' and 'after', I started victor and waited for loading to
complete, then ran PerfMetric 3 times, for 60 seconds each, and the
average of the average engine tick duration for each of those 3 sessions
was calculated.
BEFORE:  17.692 ms per vic-engine tick, average
AFTER:  16.917 ms per vic-engine tick, average
So, shaved 0.775 ms per tick on average (4.38% less time on average)

Note:  These results do not include the handful of calls in vic-engine that
  were affected, because PerfMetric is not yet set up to measure vic-anim.
  But we would expect vic-anim performance to be improved also.

* VIC-2272 mini

- const

* VIC-2272 mini PR feedback

- Include name of webviz in a const var name