CLAI-220 VIC-3480 Robot client for Jdocs service (#2971)
Similar to the existing voice and token clients, this change adds
a new client for the Jdocs service to the cloud process on the
robot. The Jdocs client starts an IPC server, where it will listen
for requests (defined in `docs.clad`) to make to the server, and
send the server's response back to whoever made the request.