VIC-5165 Move SDK anim list loading to during connection #2980
When the user requests their first animation, we don't want a delay. Instead, bury the anim list loading time during robot.connect, similar to Cozmo.

Change the animation in the tutorial to be more obvious. Unlike Cozmo, since the robot is doing freeplay before and after the tutorial, if the anim is not obvious, it is not clear whether it is part of freeplay or not.

Replace remote_control anims shown to be more Vector-specific and all working (one was failing)