VIC-5599 Prevent OTA downgrades
As a security measure, we do not want to allow the robot to be
downgraded to a prior OS version.  The prior OS version likely has
security vulnerabilities that an attacker could exploit.  The only
robots that will be allowed to downgrade are our internal "anki.dev"
robots that are running a development build.

To note, besides checking for a lower version, this change also checks:
1. The format of the os version in the manifest must match a regular
   expression for valid versions.
2. Updates from one type of build, like shipping, to a different type,
   like userdev are also considered a downgrade and not allowed.  This
   is done by checking the suffix on the version (empty for shipping,
   'd' for development and 'ud' for userdev)

If we have a production robot and we want to install the userdev build,
we will need to reboot into the factory OS and do it from there via
Chewie or mac-client.