VIC-5086 Auto-calibrate motors physical reaction (#3040)
Vector will calibrate his head and lift motor if it is detected as out of bounds

However as a caveat: it will not work when moving his head from highest, turning on calm mode, then moving it to the lowest, and turning off calm mode. This is because it considers itself to be in-air, and the behavior can't run in this scenario.