VIC-5705 Use today's UTC date when using "fake time"
Root Cause: The previous code was taking the fake UTC time and
converting it to hours and minutes and then adding that to Jan 1, 1970.
This resulted in a date/time that was not observing daylight savings
time in the tested locations.  For example, in America/Los_Angeles,
January 1 is standard time.

The Fix: Get the current UTC time and then break it out into a `struct
tm`.  Adjust the hours and minutes fields and then convert it back to
UTC seconds.

Caveats: The value returned could still be wrong if the robot's clock is
not reasonably accurate.  Given that this is for developer/QA testing,
that is acceptable.