VIC-5593 Crash fix - test which stream responses are coming from (#3186)
This crash happens when calling `Close()` on a nil stream after
an error occurs. Theoretically, this shouldn't happen: we set
`strm` to non-nil when creating one, and set it to nil after
closing it, which _should_ stop any further responses from coming.
I'm reasonably certain, given the use of `sync.Once` and the order
of operations within streams, that one stream can't send more than
one response back to the main process loop.

However, given the web of operations taking place on different
goroutines, it seems more likely that an attempt to close a stream
on the main routine doesn't stop responses from potentially coming
in due to happenings on another routine.

For that reason, this change makes it so that all responses handled
by the main process loop have the stream that sent them passed
along. Responses for which the originating stream pointer doesn't
match the current stream pointer will be logged and discarded.